+++ packages/uql-orm/src/bunSql/bunMariadb.test.ts
@@ -1,15 +1,22 @@
-import { AbstractSqlQuerierIt } from '../querier/abstractSqlQuerier-test.js';
+import { AbstractSqlQuerierPoolIt } from '../querier/abstractSqlQuerierPool-test.js';
+import { VectorQuerierIt } from '../querier/vectorQuerier-test.js';
 import { createSpec } from '../test/index.js';
+import type { BunSqlQuerier } from './bunSqlQuerier.js';
 import { BunSqlQuerierPool } from './bunSqlQuerierPool.js';
 
-class BunMariadbIt extends AbstractSqlQuerierIt {
+const url = 'mariadb://test:test@0.0.0.0:3326/test_bun_maria';
+
+class BunMariadbIt extends VectorQuerierIt {
   constructor() {
-    super(
-      new BunSqlQuerierPool({
-        url: 'mariadb://test:test@0.0.0.0:3326/test_bun_maria',
-      }),
-    );
+    super(new BunSqlQuerierPool({ url }));
+  }
+}
+
+class BunMariadbPoolIt extends AbstractSqlQuerierPoolIt<BunSqlQuerier> {
+  constructor() {
+    super(new BunSqlQuerierPool({ url }));
   }
 }
 
 createSpec(new BunMariadbIt());
+createSpec(new BunMariadbPoolIt());