+++ packages/uql-orm/src/bunSql/index.ts
@@ -1,4 +1,5 @@
 export * from './bunSqliteDialect.js';
+export * from './bunSqlCockroachDialect.js';
 export * from './bunSqlPostgresDialect.js';
 export * from './bunSqlQuerier.js';
 export * from './bunSqlQuerierPool.js';